% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-extract_scales.R
\name{extract_scales}
\alias{extract_scales}
\title{A function to extract values of remote sensing predictors at
different spatial resolutions
The resolution are resampled with an aggregation factor of 5 and 3.
For example, if using a raster at 1km res, the variables will be extracted at
~1km, ~5k and ~10km resolution.}
\usage{
extract_scales(r, p, na.rm = TRUE)
}
\arguments{
\item{r}{rasterfile of a remote sensing variable or any predictor}

\item{p}{a SpatialPointsDataFrame object}

\item{na.rm}{argument}
}
\value{
a data.frame
}
\description{
A function to extract values of remote sensing predictors at
different spatial resolutions
The resolution are resampled with an aggregation factor of 5 and 3.
For example, if using a raster at 1km res, the variables will be extracted at
~1km, ~5k and ~10km resolution.
}
