% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-fun_rich_from_suit.R
\name{spp_rich_suit}
\alias{spp_rich_suit}
\title{Species richness from suitability rasters
This function calculate species richness based on an arbitrary threshold of
0.5 of suitability rasters}
\usage{
spp_rich_suit(x, na.rm = FALSE)
}
\arguments{
\item{x}{rasterstack of habitat suitability maps}

\item{na.rm}{argument}
}
\value{
species richness raster
}
\description{
Species richness from suitability rasters
This function calculate species richness based on an arbitrary threshold of
0.5 of suitability rasters
}
