% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-fun_sum.R
\name{sum_vectors}
\alias{sum_vectors}
\title{Sum vectors > 0.5
This function calculate the sum of a vector with values larger than 0.5}
\usage{
sum_vectors(x, ...)
}
\arguments{
\item{x}{a numerical vector of habitat suitability values}

\item{...}{arguments to be passed to internal functions}
}
\value{
numeric sum
}
\description{
Sum vectors > 0.5
This function calculate the sum of a vector with values larger than 0.5
}
\examples{
suit_vector <- c(0.3, 0.7, 0.1, 0, 0.1)
sum_vectors(suit_vector)
}
