% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_cover}
\alias{tree_cover}
\title{Remote sensing data file:}
\format{
raster file
}
\source{
{created in Google Earth Engine to serve as example.}
}
\usage{
tree_cover
}
\description{
The raster file is is median percentage of annual tree cover value per pixel
estimated from Terra MODIS Vegetation Continuous Fields (VCF)
}
\keyword{datasets}
