% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-extract_scales.R
\name{extract_scales}
\alias{extract_scales}
\title{A function to extract values of remote sensing predictors at
different spatial resolutions
The resolution are resampled from the oringal raster at 1km to res
of 5k and 10km}
\usage{
extract_scales(r, p, na.rm = TRUE)
}
\arguments{
\item{r}{rasterfile of a remote sensing variable or any predictor}

\item{p}{a SpatialPointsDataFrame object}

\item{na.rm}{argument}
}
\value{
a data.frame
}
\description{
A function to extract values of remote sensing predictors at
different spatial resolutions
The resolution are resampled from the oringal raster at 1km to res
of 5k and 10km
}
